requirements.txt
openai
python-dotenv
mcp
proxies
google-genai
yfinance